Discrete-time model must inherit DISCRETE TIME to MPC/MHE #89
The default value of the self.state_discretization attribute is now set in the optimizer code and inherited from the MPC and MHE object. Moreover, the default value is correctly chosen for discrete systems.

Fixed typos in MPC and MHE code (structure3.DMStruct was incorrectly written as structre3.DMStruct and would throw an error during when raising the exception