default u0 value, stored measurement and consistency update
make_step can now be called without input (previous input is held constant)
Updated documentation. Removed redundant assertions (probably copied twice?
Measurement now refers to the current (and not next)  time step, making it consistent with the definition in the do-mpc doc. The measurement is also stored in the simulator data structure.
Fixed bug in the model class where the number of outputs would be incorrectly set to 0 if the meas_fun was not defined during model building