Integration of BONMIN to do_mpc to solve MINLP
The use of BONMIN within CASADI requires an integer list containing the information which of the opt-variables are discrete. There are three steps to complete the integration:
1) Get the integer-information from the user:
An integer-list was initialized in do_mpc.model. The element "integer=False" was added to the set_variable-method, so the user can define variables as integer-variables. An assert assures that only inputs can be integer. The name of the discrete input is stored in self.integer, which is given to the controller and the estimator.
2) Create the integer-list:
In both the controller(MPC) and the estimator(MHE) the integer-list is build from a struct with similar structure as the struct opt_x that contains the optimization variables. This is only done, if there are discrete inputs.
3) Use the right solver for the problems:
By checking, if there are discrete inputs, either BONMIN or IPOPT is chosen for solving a MINLP or a NLP, respectively.